requirements.txt
Flask>=2.0
google-cloud-speech>=2.0.0
google-cloud-translate>=3.0.0
